modules/module-09-deployment/requirements.txt
# Production Requirements
# Install with: pip install -r requirements.txt

# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Environment Variables
python-dotenv==1.0.0

# WSGI Server
gunicorn==21.2.0

# Security (if needed)
# Flask-WTF==1.2.1
# Flask-Login==0.6.3

# Logging (if needed)
# Flask-Logging==0.1.0

